pyproject.toml
[project]
name = "ecommfastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiomysql>=0.3.2",
    "alembic>=1.18.0",
    "bcrypt==4.3.0",
    "cryptography>=46.0.3",
    "fastapi[standard]>=0.128.0",
    "passlib[bcrypt]>=1.7.4",
    "python-decouple>=3.8",
    "python-jose[cryptography]>=3.5.0",
    "scalar-fastapi>=1.6.0",
    "sqlalchemy[asyncio]>=2.0.45",
]
